+++ src/components/option/index.ts
@@ -1,2 +1,0 @@
-export * from './option.interface';
-export * from './optgroup.interface';