+++ src/interfaces/optionsGatherer/optionsGatherer.interface.ts
@@ -1,3 +1,5 @@
+import {Signal} from '@angular/core';
+
 import {SelectOption} from '../selectOption/selectOption.interface';
 
 /**
@@ -8,5 +10,5 @@ export interface OptionsGatherer<TValue = unknown>
     /**
      * Array of all available options for select
      */
-    readonly availableOptions: SelectOption<TValue>[];
+    readonly availableOptions: Signal<SelectOption<TValue>[]|undefined|null>;
 }