+++ src/modules/ngSelect.module.ts
@@ -3,7 +3,7 @@ import {CommonModule} from '@angular/common';
 import {CommonDynamicModule} from '@anglr/common';
 
 import {NgSelectValuePipe} from '../pipes/ngSelectValue.pipe';
-import {NgSelectHasValuePipe} from '../pipes/ngSelectHasValue.pipe';
+import {NgSelectHasValuePipe} from '../pipes/selectHasValue/ngSelectHasValue.pipe';
 import {NgSelectControlValueAccessor} from '../misc/ngSelectControlValueAccessor.directive';
 import {BasicNormalStateComponent} from '../plugins/normalState/basic/basicNormalState.component';
 import {NoLiveSearchComponent} from '../plugins/liveSearch/no/noLiveSearch.component';
@@ -14,10 +14,10 @@ import {NoPositionerComponent} from '../plugins/positioner/no/noPositioner.compo
 import {BasicKeyboardHandlerComponent} from '../plugins/keyboardHandler/basic/basicKeyboardHandler.component';
 import {BasicValueHandlerComponent} from '../plugins/valueHandler/basic/basicValueHandler.component';
 import {OptionComponent} from '../components/option/option.component';
-import {OptGroupComponent} from '../components/option/optgroup.component';
+import {OptGroupComponent} from '../components/optionGroup/optgroup.component';
 import {NgSelectComponent} from '../components/select/select.component';
-import {NgSelectAbsoluteDirective} from '../directives/ngSelectAbsolute/ngSelectAbsolute.directive';
-import {NgSelectPlaceholderDirective} from '../directives/ngSelectPlaceholder/ngSelectPlaceholder.directive';
+import {NgSelectAbsoluteDirective} from '../directives/selectAbsolute/selectAbsolute.directive';
+import {NgSelectPlaceholderDirective} from '../directives/selectPlaceholder/selectPlaceholder.directive';
 
 /**
  * Module for select and its options