+++ src/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './selectHasValue/ngSelectHasValue.pipe';
+export * from './selectValue/ngSelectValue.pipe';