+++ src/plugins/readonlyState/readonlyState.interface.ts
@@ -1,4 +1,3 @@
-
 import {NormalStateOptions, NormalState} from '../normalState';
 
 /**